fixes after review

+++ utility-generator/src/main/java/eu/melodic/upperware/utilitygenerator/UtilityFunctionUtils.java
@@ -24,8 +24,8 @@ public class UtilityFunctionUtils { //todo: to move elsewhere
 
     }
 
-    public static double countNumberOfCores(Collection<ConfigurationElement> configuration){
-        return configuration.stream().mapToDouble(ConfigurationElement::getTotalNumberOfCores).sum();
+    public static int countNumberOfCores(Collection<ConfigurationElement> configuration) {
+        return configuration.stream().mapToInt(ConfigurationElement::getTotalNumberOfCores).sum();
     }
 
 }
